+++ Assets/Scripts/Gameplay/Spawner/HealPopUpPopUpSpawner.cs
@@ -10,11 +10,11 @@ namespace Gameplay.Spawner
             
         }
 
-        public override async void CreatePopUp(Vector3 center, int damage)
+        public override async void CreatePopUp(Vector3 center, float damage)
         {
             var popUpGameObject = await poolManager.GetObjectAsync<HealPopUpUI>();
             popUpGameObject.transform.position = center;
-            popUpGameObject.GetComponent<HealPopUpUI>().Play(damage);
+            popUpGameObject.GetComponent<HealPopUpUI>().Play((int)damage);
         }
     }
 }
